+++ fabfile/__init__.py
@@ -48,7 +48,7 @@ def code_deploy(item):
     """
     # Need warn only to allow the error to pass to celery.
     with settings(warn_only=True):
-        print('Code - Deploy\n{0}'.format(item))
+        print 'Code - Deploy\n{0}'.format(item)
         if item['meta']['code_type'] == 'library':
             code_type_dir = 'libraries'
         else:
@@ -60,8 +60,8 @@ def code_deploy(item):
             item['meta']['version'])
         create_directory_structure(code_folder)
         clone_task = clone_repo(item["git_url"], item["commit_hash"], code_folder)
-        print('Got clone response')
-        print(clone_task)
+        print 'Got clone response'
+        print clone_task
         if clone_task is True:
             if item['meta']['is_current']:
                 code_folder_current = '{0}/{1}/{2}/{2}-current'.format(
@@ -155,12 +155,6 @@ def instance_provision(instance):
         instance['sid'])
     profile = utilities.get_single_eve('code', instance['code']['profile'])
     profile_name = profile['meta']['name']
-
-    try:
-        result_create_database = execute(create_database, instance=instance)
-    except FabricException:
-        print 'Database creation failed.'
-        return result_create_database
 
     try:
         result_create_dir_structure = execute(
@@ -487,7 +481,7 @@ def command_run(instance, command):
     :param command: Command to run
     :return:
     """
-    print('Command - {0}\n{1}'.format(instance['sid'], command))
+    print 'Command - {0}\n{1}'.format(site['sid'], command)
     web_directory = '{0}/{1}/{2}'.format(
         instances_web_root,
         instance['type'],
@@ -504,17 +498,18 @@ def update_database(instance):
     :param instance: Instance to run command on
     :return:
     """
-    print('Database Update | {0}'.format(instance))
+    print 'Database Update\n{0}'.format(site)
     code_directory_sid = '{0}/{1}/{1}'.format(instances_code_root, instance['sid'])
     with cd(code_directory_sid):
-        print('Running database updates.')
-        run("drush updb -y")
+        print 'Running database updates.'
+        run('drush updb -y')
 
 
 @roles('webserver_single')
 def registry_rebuild(instance):
     """
-    Run a drush rr
+    Run a drush rr and drush cc drush. 
+    Drush command cache clear is a workaround, see #306.
 
     :param instance: Instance to run command on
     :return:
@@ -522,18 +517,18 @@ def registry_rebuild(instance):
     print 'Drush registry rebuild | {0}'.format(instance)
     code_directory_sid = '{0}/{1}/{1}'.format(instances_code_root, instance['sid'])
     with cd(code_directory_sid):
-        run("drush rr")
+        run('drush rr; drush cc drush;')
 
 
 @roles('webservers')
 def clear_apc():
-    run("wget -q -O - http://localhost/sysadmintools/apc/clearapc.php")
+    run('wget -q -O - http://localhost/sysadmintools/apc/clearapc.php;')
 
 
 def drush_cache_clear(sid):
     code_directory_current = '{0}/{1}/current'.format(instances_code_root, sid)
     with cd(code_directory_current):
-        run("drush cc all")
+        run('drush cc all')
 
 
 @roles('webservers')
@@ -707,7 +702,7 @@ def create_settings_files(instance, profile_name):
                     use_jinja=True,
                     template_dir=template_dir,
                     backup=False,
-                    mode='0664')
+                    mode='0644')
 
     settings_variables = {
         'profile': profile_name,
@@ -726,7 +721,7 @@ def create_settings_files(instance, profile_name):
                     use_jinja=True,
                     template_dir=template_dir,
                     backup=False,
-                    mode='0664')
+                    mode='0644')
 
     local_post_settings_variables = {
         'sid': sid,
@@ -745,14 +740,14 @@ def create_settings_files(instance, profile_name):
                     use_jinja=True,
                     template_dir=template_dir,
                     backup=False,
-                    mode='0664')
+                    mode='0644')
 
 
 @runs_once
 def install_instance(profile_name, code_directory_current):
     with cd(code_directory_current):
         run('drush site-install -y {0}'.format(profile_name))
-        run('drush rr')
+        run('drush rr; drush cc drush')
 
 
 def clone_repo(git_url, checkout_item, destination):